requirements.txt

opencv-python
moviepy
numpy
google-generativeai
python-dotenv
pyautogui
pygetwindow
Pillow
PySocks
